Added totalWinsWithPositiveScore Criteria for Commendations
Meta Missions such as the Hangar mission in the 40k mod can be triggered at any time (finalDestination: true) and will result in a win (chronoTrigger: 2). On the other hand it will always generate a score of 0. To avoid that missions such as the hangar one can be used to accumulate number of totalWins for commendations, getWinTotalWithPositiveScore with an additional check for score > 0 was introduced.